Move common dev code out of main package directory
