+++ packages/bulma/tsconfig.json
@@ -4,5 +4,5 @@
     "outDir": "dist/types",
     "skipLibCheck": true
   },
-  "include": ["./src/"]
+  "include": ["./src/", "./dev/"]
 }
