+++ packages/chakra/dev/main.tsx
@@ -2,8 +2,8 @@ import { ChakraProvider, extendTheme } from '@chakra-ui/react';
 import * as React from 'react';
 import { createRoot } from 'react-dom/client';
 import { QueryBuilder } from 'react-querybuilder';
-import { DevLayout, useDevApp } from 'react-querybuilder/dev/index.js';
-import 'react-querybuilder/dist/query-builder.scss';
+import { DevLayout, useDevApp } from '@dev';
+import 'react-querybuilder/src/query-builder.scss';
 import { QueryBuilderChakra } from '../src';
 import './styles.scss';
