+++ packages/native/vite.config.mts
@@ -1,10 +1,12 @@
 import vitePluginReact from '@vitejs/plugin-react-swc';
+import path from 'node:path';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
   resolve: {
     alias: {
       'react-native': 'react-native-web',
+      'react-querybuilder': path.resolve(import.meta.dir, '../react-querybuilder'),
     },
   },
   plugins: [vitePluginReact()],
