+++ utils/dev/musicalInstruments.ts
@@ -1,5 +1,5 @@
 // Adapted from https://en.wikipedia.org/wiki/List_of_musical_instruments
-import type { OptionGroup } from '../src';
+import type { OptionGroup } from 'react-querybuilder';
 
 export const musicalInstruments: OptionGroup[] = [
   {
