+++ utils/dev/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "dev",
+  "version": "0.0.0",
+  "private": true,
+  "exports": {
+    ".": {
+      "import": "./index.ts",
+      "types": "./index.ts"
+    }
+  }
+}
