+++ utils/dev/styles.scss
@@ -1,4 +1,4 @@
-@import '../src/query-builder.scss';
+@import 'react-querybuilder/src/query-builder.scss';
 
 body {
   margin: 0;
